changed structure of server code
audio gets reversed when posted to server instead of when getting. The reason is so the client doesn't waste resources on reversing the audio, every time they have to get the audio.  By reversing the audio on post requests, the reversed audio is saved on the server, and does not need extra resources to be spent on it.